MapData/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
geoalchemy2>=0.9.0
python-dotenv>=0.19.0
alembic>=1.7.0
pydantic>=1.8.2 